requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
albumentations=1.3.1=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
filelock=3.12.2=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.42.1=pypi_0
fsspec=2023.6.0=pypi_0
gast=0.5.4=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.57.0=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
joblib=1.3.2=pypi_0
keras=3.1.1=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libclang=18.1.1=pypi_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgomp=13.1.0=he5830b7_0
libnsl=2.0.0=h7f98852_0
libsqlite=3.42.0=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=hd590300_5
markdown=3.4.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.3.2=pypi_0
namex=0.0.7=pypi_0
ncurses=6.4=hcb278e6_0
networkx=3.1=pypi_0
numpy=1.25.2=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.76=pypi_0
opencv-python-headless=4.8.0.76=pypi_0
openssl=3.1.2=hd590300_0
opt-einsum=3.3.0=pypi_0
optree=0.11.0=pypi_0
packaging=23.1=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=pyhd8ed1ab_0
protobuf=4.24.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.12=hd12c33a_0_cpython
python-dateutil=2.8.2=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
setuptools=68.1.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.8.12=pypi_0
timm=0.9.5=pypi_0
tk=8.6.12=h27826a3_0
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.16=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
xz=5.2.6=h166bdaf_0
yacs=0.1.8=pypi_0
